Fix flags compiling errors

+++ src/sparse_hopnet/includes/SparseHN.hpp
@@ -50,7 +50,7 @@ class SparseHN
 
         // Inference
         float   token_energy(std::vector<t_iclamped> const &clamped,
-                                int i,
+                                size_t i,
                                 size_t seq_len);
         float   seq_energy(std::vector<t_iclamped> const &clamped);
         float   eval(std::vector<t_iclamped>  const &clamped, size_t id);
